vyuha/backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
supabase==2.0.2
python-dotenv==1.0.0
openpyxl==3.1.2
pydantic>=2.7.0
httpx==0.24.1
bcrypt==4.1.3
PyJWT==2.9.0
starlette>=0.27.0
requests>=2.31.0
pytest>=7.4.0
python-multipart>=0.0.9
